Added ansible.cfg for automatic hostfile detection Uncommented vars to version control ansible-vault encrypted file Updated teardown of sim

+++ simulation.sh
@@ -7,11 +7,20 @@ SERVERS=(leela bigbertha boole lovelace)
 if [[ $1 == 'teardown' ]]; then
     for server in ${SERVERS[@]}; do
         ip=$(lxc list --format json $server | jq -r .[0].state.network.eth0.addresses[0].address)
-        ssh-keygen -R $ip > /dev/null
-        printf '.'
-        lxc stop $server
-        printf '.'
-        lxc rm $server
+        $(lxc ls | grep $server) &> /dev/null
+        exists=$?
+        
+        if [[ exists == 0 ]]; then
+            state=$(lxc list --format json $server | jq -r .[0].status)
+            (ssh-keygen -R $ip) &> /dev/null
+
+            printf '.'
+            if [[ $state != "Stopped" ]]; then
+                lxc stop $server
+            fi
+            printf '.'
+            lxc rm $server
+        fi
     done
     echo 'Done!'
     exit 0
